// apps/pwabuilder-chromium-extension/src/interfaces/manifest.ts
export interface TestResult {
  infoString: string;
  result: boolean;
  category: string;
  description?: string;
}